feat(scalars): joins single and multiple select components into a single select component
