+++ packages/design-system/src/scalars/components/fragments/badge/badge.test.tsx
@@ -1,0 +1,10 @@
+import { describe, expect, it } from "vitest";
+import { render } from "@testing-library/react";
+import { Badge } from "./badge";
+
+describe("Badge Component", () => {
+  it("should match snapshot", () => {
+    const { asFragment } = render(<Badge />);
+    expect(asFragment()).toMatchSnapshot();
+  });
+});
