+++ packages/design-system/src/scalars/components/fragments/badge/index.ts
@@ -1,0 +1,1 @@
+export * from "./badge";
