+++ packages/design-system/src/scalars/components/fragments/button/button.test.tsx
@@ -1,0 +1,10 @@
+import { describe, expect, it } from "vitest";
+import { render } from "@testing-library/react";
+import { Button } from "./button";
+
+describe("Button Component", () => {
+  it("should match snapshot", () => {
+    const { asFragment } = render(<Button />);
+    expect(asFragment()).toMatchSnapshot();
+  });
+});
