+++ packages/design-system/src/scalars/components/fragments/button/index.ts
@@ -1,0 +1,1 @@
+export * from "./button";
