+++ packages/design-system/src/scalars/components/fragments/command/command.test.tsx
@@ -1,0 +1,10 @@
+import { describe, expect, it } from "vitest";
+import { render } from "@testing-library/react";
+import { Command } from "./command";
+
+describe("Command Component", () => {
+  it("should match snapshot", () => {
+    const { asFragment } = render(<Command />);
+    expect(asFragment()).toMatchSnapshot();
+  });
+});
