+++ packages/design-system/src/scalars/components/fragments/dialog/index.ts
@@ -1,0 +1,1 @@
+export * from "./dialog";
