--- packages/design-system/src/scalars/components/fragments/multi-select-field/index.ts
@@ -1,1 +1,0 @@
-export * from "./multi-select-field";
