+++ packages/design-system/src/scalars/components/fragments/popover/index.ts
@@ -1,0 +1,1 @@
+export * from "./popover";
