+++ packages/design-system/src/scalars/components/fragments/popover/popover.test.tsx
@@ -1,0 +1,10 @@
+import { describe, expect, it } from "vitest";
+import { render } from "@testing-library/react";
+import { Popover } from "./popover";
+
+describe("Popover Component", () => {
+  it("should match snapshot", () => {
+    const { asFragment } = render(<Popover />);
+    expect(asFragment()).toMatchSnapshot();
+  });
+});
