+++ packages/design-system/src/scalars/components/fragments/select-field/select-field.test.tsx
@@ -1,9 +1,9 @@
 import { describe, expect, it, vi } from "vitest";
 import { render, screen } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
-import { SingleSelectField } from "./single-select-field";
+import { SelectField } from "./select-field";
 
-describe("SingleSelectField Component", () => {
+describe("SelectField Component", () => {
   const defaultOptions = [
     { label: "Option 1", value: "1" },
     { label: "Option 2", value: "2" },
@@ -14,20 +14,20 @@ describe("SingleSelectField Component", () => {
   // Basic Rendering Tests
   it("should match snapshot", () => {
     const { asFragment } = render(
-      <SingleSelectField name="single-select" options={defaultOptions} />,
+      <SelectField name="select" options={defaultOptions} />,
     );
     expect(asFragment()).toMatchSnapshot();
   });
 
   it("should render with default placeholder", () => {
-    render(<SingleSelectField name="single-select" options={defaultOptions} />);
+    render(<SelectField name="select" options={defaultOptions} />);
     expect(screen.getByText("Select an option")).toBeInTheDocument();
   });
 
   it("should render with custom placeholder", () => {
     render(
-      <SingleSelectField
-        name="single-select"
+      <SelectField
+        name="select"
         options={defaultOptions}
         placeholder="Custom placeholder"
       />,
@@ -38,19 +38,15 @@ describe("SingleSelectField Component", () => {
   // Label Tests
   it("should render with label", () => {
     render(
-      <SingleSelectField
-        name="single-select"
-        options={defaultOptions}
-        label="Test Label"
-      />,
+      <SelectField name="select" options={defaultOptions} label="Test Label" />,
     );
     expect(screen.getByText("Test Label")).toBeInTheDocument();
   });
 
   it("should show required indicator when required", () => {
     render(
-      <SingleSelectField
-        name="single-select"
+      <SelectField
+        name="select"
         options={defaultOptions}
         label="Test Label"
         required
@@ -66,8 +62,8 @@ describe("SingleSelectField Component", () => {
     const user = userEvent.setup();
 
     render(
-      <SingleSelectField
-        name="single-select"
+      <SelectField
+        name="select"
         options={defaultOptions}
         onChange={onChangeMock}
       />,
@@ -85,8 +81,8 @@ describe("SingleSelectField Component", () => {
     const user = userEvent.setup();
 
     render(
-      <SingleSelectField
-        name="single-select"
+      <SelectField
+        name="select"
         options={defaultOptions}
         onChange={onChangeMock}
       />,
@@ -101,13 +97,7 @@ describe("SingleSelectField Component", () => {
   // Search Functionality Tests
   it("should show search input when searchable is true", async () => {
     const user = userEvent.setup();
-    render(
-      <SingleSelectField
-        name="single-select"
-        options={defaultOptions}
-        searchable
-      />,
-    );
+    render(<SelectField name="select" options={defaultOptions} searchable />);
 
     await user.click(screen.getByRole("combobox"));
     expect(screen.getByPlaceholderText("Search...")).toBeInTheDocument();
@@ -115,13 +105,7 @@ describe("SingleSelectField Component", () => {
 
   it("should filter options based on search input", async () => {
     const user = userEvent.setup();
-    render(
-      <SingleSelectField
-        name="single-select"
-        options={defaultOptions}
-        searchable
-      />,
-    );
+    render(<SelectField name="select" options={defaultOptions} searchable />);
 
     await user.click(screen.getByRole("combobox"));
     await user.type(screen.getByPlaceholderText("Search..."), "Option 1");
@@ -133,8 +117,8 @@ describe("SingleSelectField Component", () => {
   // // Validation and Error Handling Tests
   it("should display error messages", () => {
     render(
-      <SingleSelectField
-        name="single-select"
+      <SelectField
+        name="select"
         options={defaultOptions}
         errors={["This field is required"]}
       />,
@@ -144,8 +128,8 @@ describe("SingleSelectField Component", () => {
 
   it("should display warning messages", () => {
     render(
-      <SingleSelectField
-        name="single-select"
+      <SelectField
+        name="select"
         options={defaultOptions}
         warnings={["Please review your selection"]}
       />,
@@ -158,7 +142,7 @@ describe("SingleSelectField Component", () => {
   // // Keyboard Navigation Tests
   it("should handle keyboard navigation", async () => {
     const user = userEvent.setup();
-    render(<SingleSelectField name="single-select" options={defaultOptions} />);
+    render(<SelectField name="select" options={defaultOptions} />);
 
     screen.getByRole("combobox").focus(); // Focus the combobox
 
@@ -171,37 +155,20 @@ describe("SingleSelectField Component", () => {
 
   // // Edge Cases
   it("should handle empty options array", () => {
-    render(<SingleSelectField name="single-select" options={[]} />);
+    render(<SelectField name="select" options={[]} />);
     expect(screen.getByText("Select an option")).toBeInTheDocument();
   });
 
-  it("should handle controlled value updates", () => {
-    const { rerender } = render(
-      <SingleSelectField
-        name="single-select"
-        options={defaultOptions}
-        value="1"
-      />,
-    );
-
-    expect(screen.getByText("Option 1")).toBeInTheDocument();
-
-    rerender(
-      <SingleSelectField
-        name="single-select"
-        options={defaultOptions}
-        value="2"
-      />,
-    );
-
+  it("should handle value when provided", () => {
+    render(<SelectField name="select" options={defaultOptions} value="2" />);
     expect(screen.getByText("Option 2")).toBeInTheDocument();
   });
 
   // Accessibility Tests
   it("should have correct ARIA attributes", () => {
     render(
-      <SingleSelectField
-        name="single-select"
+      <SelectField
+        name="select"
         options={defaultOptions}
         label="Test Label"
         required
@@ -218,15 +185,30 @@ describe("SingleSelectField Component", () => {
   // Modal Behavior Tests
   it("should render as modal when asModal is true", async () => {
     const user = userEvent.setup();
+    render(<SelectField name="select" options={defaultOptions} asModal />);
+
+    await user.click(screen.getByRole("combobox"));
+    expect(screen.getByRole("dialog")).toBeInTheDocument();
+  });
+
+  // Multiple Selection Tests
+  it("should allow multiple selections when multiple is true", async () => {
+    const onChangeMock = vi.fn();
+    const user = userEvent.setup();
+
     render(
-      <SingleSelectField
-        name="single-select"
+      <SelectField
+        name="select"
         options={defaultOptions}
-        asModal
+        onChange={onChangeMock}
+        multiple
       />,
     );
 
     await user.click(screen.getByRole("combobox"));
-    expect(screen.getByRole("dialog")).toBeInTheDocument();
+    await user.click(screen.getByText("Option 1"));
+    expect(onChangeMock).toHaveBeenCalledWith(["1"]);
+    await user.click(screen.getByText("Option 2"));
+    expect(onChangeMock).toHaveBeenLastCalledWith(["1", "2"]);
   });
 });
