+++ packages/design-system/src/scalars/components/fragments/separator/index.ts
@@ -1,0 +1,1 @@
+export * from "./separator";
