+++ packages/design-system/src/scalars/components/fragments/separator/separator.tsx
