--- packages/design-system/src/scalars/components/fragments/single-select-field/index.ts
@@ -1,1 +1,0 @@
-export * from "./single-select-field";
